+++ Service/StaRepo/src/main/java/com/hzgc/service/starepo/bean/PrisonSearchOpts.java
@@ -1,4 +1,4 @@
-package com.hzgc.service.starepo.object;
+package com.hzgc.service.starepo.bean;
 
 import java.io.Serializable;
 import java.util.List;
