+++ Service/StaRepo/src/main/java/com/hzgc/service/starepo/bean/SubQueryOpts.java
@@ -1,4 +1,4 @@
-package com.hzgc.service.starepo.object;
+package com.hzgc.service.starepo.bean;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
