+++ Service/StaRepo/src/main/java/com/hzgc/service/starepo/controller/ObjectTypeController.java
@@ -1,9 +1,10 @@
 package com.hzgc.service.starepo.controller;
 
+import com.alibaba.druid.pool.DruidDataSource;
 import com.hzgc.common.service.BigDataPath;
 import com.hzgc.common.service.ResponseResult;
-import com.hzgc.service.starepo.service.ObjectTypeServiceImpl;
-import com.hzgc.service.starepo.vo.ObjectTypeVO;
+import com.hzgc.service.starepo.service.ObjectTypeService;
+import com.hzgc.service.starepo.bean.ObjectType;
 import io.swagger.annotations.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.netflix.feign.FeignClient;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.annotation.Resource;
 import java.util.List;
 import java.util.Map;
 
@@ -22,18 +24,18 @@ import java.util.Map;
 public class ObjectTypeController {
 
     @Autowired
-    private ObjectTypeServiceImpl objectTypeServiceImpl;
+    private ObjectTypeService objectTypeService;
 
     @ApiOperation(value = "添加对象类型", response = Byte.class)
     @ApiResponses(
             {@ApiResponse(code = 200, message = "successful response")})
     @RequestMapping(value = BigDataPath.TYPE_ADD, method = RequestMethod.POST)
-    public ResponseResult<Boolean> addObjectType(@RequestBody @ApiParam(value = "对象类型封装类") ObjectTypeVO vo) {
+    public ResponseResult<Boolean> addObjectType(@RequestBody @ApiParam(value = "对象类型封装类") ObjectType vo) {
         if(vo == null || vo.getName() == null || "".equals(vo.getName())){
             return null;
         }
-
-        boolean success = objectTypeServiceImpl.addObjectType(vo.getName(), vo.getCreator(), vo.getRemark());
+        System.out.println(vo.toString());
+        boolean success = objectTypeService.addObjectType(vo.getName(), vo.getCreator(), vo.getRemark());
         return ResponseResult.init(success);
     }
 
@@ -41,12 +43,12 @@ public class ObjectTypeController {
     @ApiResponses(
             {@ApiResponse(code = 200, message = "successful response")})
     @RequestMapping(value = BigDataPath.TYPE_DELETE, method = RequestMethod.POST)
-    public ResponseResult<Boolean> deleteObjectType(@RequestBody @ApiParam(value = "对象类型封装类") ObjectTypeVO vo) {
+    public ResponseResult<Boolean> deleteObjectType(@RequestBody @ApiParam(value = "对象类型封装类") ObjectType vo) {
         if(vo == null || vo.getId() == null || "".equals(vo.getId())){
             return null;
         }
 
-        boolean success = objectTypeServiceImpl.deleteObjectType(vo.getId());
+        boolean success = objectTypeService.deleteObjectType(vo.getId());
         return ResponseResult.init(success);
     }
 
@@ -54,12 +56,12 @@ public class ObjectTypeController {
     @ApiResponses(
             {@ApiResponse(code = 200, message = "successful response")})
     @RequestMapping(value = BigDataPath.TYPE_UPDATE, method = RequestMethod.POST)
-    public ResponseResult<Boolean> updateObjectType(@RequestBody @ApiParam(value = "对象类型封装类") ObjectTypeVO vo) {
+    public ResponseResult<Boolean> updateObjectType(@RequestBody @ApiParam(value = "对象类型封装类") ObjectType vo) {
         if(vo == null || vo.getId() == null || vo.getName() == null || "".equals(vo.getId()) || "".equals(vo.getName())){
             return null;
         }
 
-        boolean success = objectTypeServiceImpl.updateObjectType(vo.getId(), vo.getName(), vo.getCreator(), vo.getRemark());
+        boolean success = objectTypeService.updateObjectType(vo.getId(), vo.getName(), vo.getCreator(), vo.getRemark());
         return ResponseResult.init(success);
     }
 
@@ -67,12 +69,12 @@ public class ObjectTypeController {
     @ApiResponses(
             {@ApiResponse(code = 200, message = "successful response")})
     @RequestMapping(value = BigDataPath.TYPE_SEARCH, method = RequestMethod.POST)
-    public ResponseResult<List> searchObjectType(@RequestBody @ApiParam(value = "对象类型封装类") ObjectTypeVO vo) {
+    public ResponseResult<List> searchObjectType(@RequestBody @ApiParam(value = "对象类型封装类") ObjectType vo) {
         if(vo == null || vo.getPageIndex() == 0 || vo.getPageSize() == 0){
             return null;
         }
 
-        List<Map<String, String>> success = objectTypeServiceImpl.searchObjectType(vo.getName(), vo.getPageIndex(), vo.getPageSize());
+        List<Map<String, String>> success = objectTypeService.searchObjectType(vo.getName(), vo.getPageIndex(), vo.getPageSize());
         return ResponseResult.init(success);
     }
 }
