+++ Service/StaRepo/src/main/java/com/hzgc/service/starepo/dao/ObjectTypeDao.java
@@ -1,18 +1,22 @@
-package com.hzgc.service.starepo.service;
+package com.hzgc.service.starepo.dao;
 
-import com.hzgc.service.starepo.util.PhoenixJDBCHelper;
-import org.apache.log4j.Logger;
-import org.springframework.stereotype.Service;
 import com.hzgc.common.service.table.column.ObjectInfoTable;
-
+import org.apache.log4j.Logger;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.support.rowset.SqlRowSet;
+import org.springframework.stereotype.Repository;
+import javax.annotation.Resource;
 import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.*;
 
-@Service
-public class ObjectTypeServiceImpl implements ObjectTypeService {
-    private static Logger LOG = Logger.getLogger(ObjectInfoHandlerImpl.class);
+@Repository
+public class ObjectTypeDao {
+
+    private static Logger LOG = Logger.getLogger(ObjectTypeDao.class);
+
+    @Resource(name = "phoenixJdbcTemplate")
+    private JdbcTemplate jdbcTemplate;
 
     /**
      * 添加objectType
@@ -21,13 +25,12 @@ public class ObjectTypeServiceImpl implements ObjectTypeService {
      * @param remark
      * @return
      */
-    @Override
     public boolean addObjectType(String name, String creator, String remark) {
         LOG.info("objectType" + name);
         if(name == null || "".equals(name)){
+            LOG.info("name is null");
             return false;
         }
-        java.sql.Connection conn = null;
         long start = System.currentTimeMillis();
         String uuid = UUID.randomUUID().toString().replace("-", "");
         String typeId = "type_" + System.currentTimeMillis() + uuid.substring(0, 8);
@@ -38,22 +41,11 @@ public class ObjectTypeServiceImpl implements ObjectTypeService {
                 + ObjectInfoTable.TYPE_COLF + "." + ObjectInfoTable.TYPE_ADD_TIME
                 +") values (?,?,?,?,?)";
         LOG.info("sql:" + sql);
-        PreparedStatement pstm = null;
         try {
-            conn = PhoenixJDBCHelper.getInstance().getConnection();
-            pstm = conn.prepareStatement(sql);
-            pstm.setString(1, typeId);
-            pstm.setString(2, name);
-            pstm.setString(3, creator);
-            pstm.setString(4, remark);
-            pstm.setTimestamp(5,new java.sql.Timestamp(System.currentTimeMillis()));
-            pstm.executeUpdate();
-            conn.commit();
-        } catch (SQLException e) {
+            jdbcTemplate.update(sql,typeId,name,creator,remark,new java.sql.Timestamp(System.currentTimeMillis()));
+        } catch (Exception e) {
             LOG.error(e.getMessage());
             return false;
-        } finally {
-            PhoenixJDBCHelper.closeConnection(null, pstm, null);
         }
         LOG.info("添加一条数据到静态库花费时间： " + (System.currentTimeMillis() - start));
         return true;
@@ -64,7 +56,6 @@ public class ObjectTypeServiceImpl implements ObjectTypeService {
      * @param id
      * @return
      */
-    @Override
     public boolean deleteObjectType(String id) {
         LOG.info("rowkey to delete : " + id);
 
@@ -72,31 +63,19 @@ public class ObjectTypeServiceImpl implements ObjectTypeService {
             return false;
         }
         long start = System.currentTimeMillis();
-        java.sql.Connection conn = null;
         String sql1 = "select id from objectinfo where " + ObjectInfoTable.PKEY + " = ?";
         String sql2 = "delete from objectinfo where " + ObjectInfoTable.ROWKEY + " = ?";
-        PreparedStatement pstm = null;
         try {
-            conn = PhoenixJDBCHelper.getInstance().getConnection();
-            pstm = conn.prepareStatement(sql1);
-            pstm.setString(1,id);
-            ResultSet resultSet = pstm.executeQuery();
-            while (resultSet.next()){
-                String rowkey = resultSet.getString(ObjectInfoTable.ROWKEY);
+            SqlRowSet sqlRowSet = jdbcTemplate.queryForRowSet(sql1,id);
+            while (sqlRowSet.next()){
+                String rowkey = sqlRowSet.getString(ObjectInfoTable.ROWKEY);
                 LOG.info("存在该类型的对象，不能删除");
-                PhoenixJDBCHelper.closeConnection(null, pstm, null);
                 return false;
             }
-            pstm = conn.prepareStatement(sql2);
-            pstm.setString(1, id);
-            pstm.executeUpdate();
-            conn.commit();
-
-        } catch (SQLException e) {
+            jdbcTemplate.update(sql2,id);
+        } catch (Exception e) {
             e.printStackTrace();
             return false;
-        } finally {
-            PhoenixJDBCHelper.closeConnection(null, pstm, null);
         }
         LOG.info("删除静态信息库的" + id + "数据花费时间： " + (System.currentTimeMillis() - start));
         return true;
@@ -110,7 +89,6 @@ public class ObjectTypeServiceImpl implements ObjectTypeService {
      * @param remark
      * @return
      */
-    @Override
     public boolean updateObjectType(String  id, String name, String creator, String remark) {
         LOG.info("objectType" + id + " : " + name);
         if(id == null || "".equals(id)){
@@ -119,7 +97,6 @@ public class ObjectTypeServiceImpl implements ObjectTypeService {
         if(name == null || "".equals(name)){
             return false;
         }
-        java.sql.Connection conn = null;
         long start = System.currentTimeMillis();
         String sql = "upsert into objectinfo(" + ObjectInfoTable.ROWKEY + ", "
                 + ObjectInfoTable.TYPE_COLF + "." + ObjectInfoTable.TYPE_NAME + ", "
@@ -127,22 +104,11 @@ public class ObjectTypeServiceImpl implements ObjectTypeService {
                 + ObjectInfoTable.TYPE_COLF + "." + ObjectInfoTable.TYPE_REMARK + ", "
                 + ObjectInfoTable.TYPE_COLF + "." + ObjectInfoTable.TYPE_UPDATE_TIME
                 +") values (?,?,?,?,?)";
-        PreparedStatement pstm = null;
         try {
-            conn = PhoenixJDBCHelper.getInstance().getConnection();
-            pstm = conn.prepareStatement(sql);
-            pstm.setString(1, id);
-            pstm.setString(2, name);
-            pstm.setString(3, creator);
-            pstm.setString(4, remark);
-            pstm.setTimestamp(5,new java.sql.Timestamp(System.currentTimeMillis()));
-            pstm.executeUpdate();
-            conn.commit();
-        } catch (SQLException e) {
+            jdbcTemplate.update(sql,id,name,creator,remark,new java.sql.Timestamp(System.currentTimeMillis()));
+        } catch (Exception e) {
             LOG.error(e.getMessage());
             return false;
-        } finally {
-            PhoenixJDBCHelper.closeConnection(null, pstm, null);
         }
         LOG.info("添加一条数据到静态库花费时间： " + (System.currentTimeMillis() - start));
         return true;
@@ -155,7 +121,6 @@ public class ObjectTypeServiceImpl implements ObjectTypeService {
      * @param pageSize
      * @return
      */
-    @Override
     public List<Map<String, String>> searchObjectType(String name, int pageIndex, int pageSize) {
         if(pageIndex == 0){
             pageIndex = 1;
@@ -165,7 +130,7 @@ public class ObjectTypeServiceImpl implements ObjectTypeService {
         }
         List<Map<String, String>> result = null;
         java.sql.Connection conn = null;
-        ResultSet resultSet = null;
+        SqlRowSet sqlRowSet = null;
         PreparedStatement pstm = null;
         StringBuilder sql = new StringBuilder("select " + ObjectInfoTable.ROWKEY + ", "
                 + ObjectInfoTable.TYPE_COLF + "." + ObjectInfoTable.TYPE_NAME + ", "
@@ -179,46 +144,42 @@ public class ObjectTypeServiceImpl implements ObjectTypeService {
         }
         sql.append(" LIMIT " + pageSize);
         try {
-            conn = PhoenixJDBCHelper.getInstance().getConnection();
-            pstm = conn.prepareStatement(sql.toString());
             String startRow = "a";
             if(pageIndex == 1){
-                resultSet = getDate(pstm, startRow);
-                result = getResult(resultSet);
+                sqlRowSet = getDate(sql.toString(), startRow);
+                result = getResult(sqlRowSet);
             }else{
                 for(int i = 1;i< pageIndex; i++){
-                    resultSet = getDate(pstm, startRow);
-                    startRow = getLastRowkey(resultSet);
+                    sqlRowSet = getDate(sql.toString(), startRow);
+                    startRow = getLastRowkey(sqlRowSet);
                 }
-                resultSet = getDate(pstm, startRow);
-                result = getResult(resultSet);
+                sqlRowSet = getDate(sql.toString(), startRow);
+                result = getResult(sqlRowSet);
             }
         } catch (SQLException e) {
             LOG.error(e.getMessage());
             return null;
-        } finally {
-            PhoenixJDBCHelper.closeConnection(null, pstm, null);
         }
         return result;
     }
 
     /**
      * 从ResultSet中取出数据
-     * @param resultSet
+     * @param sqlRowSet
      * @return
      * @throws SQLException
      */
-    private List<Map<String, String>> getResult(ResultSet resultSet) throws SQLException {
+    private List<Map<String, String>> getResult(SqlRowSet sqlRowSet) throws SQLException {
         List<Map<String, String>> result = new ArrayList<>();
-        while (resultSet.next()){
+        while (sqlRowSet.next()){
             Map<String, String> map = new HashMap<>();
-            map.put(ObjectInfoTable.ROWKEY, resultSet.getString(ObjectInfoTable.ROWKEY));
+            map.put(ObjectInfoTable.ROWKEY, sqlRowSet.getString(ObjectInfoTable.ROWKEY));
             map.put(ObjectInfoTable.TYPE_NAME,
-                    resultSet.getString(ObjectInfoTable.TYPE_NAME));
+                    sqlRowSet.getString(ObjectInfoTable.TYPE_NAME));
             map.put(ObjectInfoTable.TYPE_CREATOR,
-                    resultSet.getString(ObjectInfoTable.TYPE_CREATOR));
+                    sqlRowSet.getString(ObjectInfoTable.TYPE_CREATOR));
             map.put(ObjectInfoTable.TYPE_REMARK,
-                    resultSet.getString(ObjectInfoTable.TYPE_REMARK));
+                    sqlRowSet.getString(ObjectInfoTable.TYPE_REMARK));
             result.add(map);
         }
         return result;
@@ -226,28 +187,28 @@ public class ObjectTypeServiceImpl implements ObjectTypeService {
 
     /**
      * 获取一页数据
-     * @param pstm
+     * @param sql
      * @param startRow
      * @return
      * @throws SQLException
      */
-    private ResultSet getDate(PreparedStatement pstm , String startRow) throws SQLException {
-        pstm.setString(1, startRow);
-        ResultSet resultSet = pstm.executeQuery();
-        return resultSet;
+    private SqlRowSet getDate(String sql , String startRow) throws SQLException {
+        SqlRowSet sqlRowSet = jdbcTemplate.queryForRowSet(sql,startRow);
+        return sqlRowSet;
     }
 
     /**
      * 取得最后一行Rowkey
-     * @param resultSet
+     * @param sqlRowSet
      * @return
      * @throws SQLException
      */
-    private String getLastRowkey(ResultSet resultSet) throws SQLException {
+    private String getLastRowkey(SqlRowSet sqlRowSet) throws SQLException {
         String lastRowKey = null;
-        while(resultSet.next()){
-            lastRowKey = resultSet.getString(ObjectInfoTable.ROWKEY);
+        while(sqlRowSet.next()){
+            lastRowKey = sqlRowSet.getString(ObjectInfoTable.ROWKEY);
         }
         return lastRowKey;
     }
+
 }
