+++ Service/StaRepo/src/main/resources/starepoApplication.properties
@@ -7,7 +7,7 @@ eureka.client.fetch-registry=false
 
 ##### 连接池相关配置 #####
 phoenix.enable=true
-phoenix.url=jdbc:phoenix:172.16.0.20
+phoenix.url=jdbc:phoenix:172.18.18.100:2181
 phoenix.type=com.alibaba.druid.pool.DruidDataSource
 phoenix.driver-class-name=org.apache.phoenix.jdbc.PhoenixDriver
 ##phoenix的用户名默认为空
