+++ Service/StaRepo/src/test/java/com/hzgc/testobjecttype/ObjectTypeTest.java
@@ -1,0 +1,11 @@
+package com.hzgc.testobjecttype;
+
+import com.hzgc.service.starepo.service.ObjectTypeService;
+
+public class ObjectTypeTest {
+
+    public void addType(){
+        ObjectTypeService objectTypeService = new ObjectTypeService();
+        
+    }
+}
