+++ Service/StaRepo/src/test/resources/core-site.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?xml-stylesheet type="text/xsl" href="configuration.xsl"?>
+<!--
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License. See accompanying LICENSE file.
+-->
+
+<!-- Put site-specific property overrides in this file. -->
+
+<configuration>
+  
+    <!-- 指定hdfs的nameservice为ns1，可以简单地理解为集群的名字 -->
+    <property>
+        <name>fs.defaultFS</name>
+        <value>hdfs://hzgc/</value>
+    </property>
+
+    <!-- 指定hadoop临时目录 -->
+    <property>
+        <name>hadoop.tmp.dir</name>
+        <value>/opt/hzgc/bigdata/Hadoop/hadoop/tmp</value>
+    </property>
+	
+    <!-- 指定zookeeper地址 -->
+    <property>
+	<name>ha.zookeeper.quorum</name>
+	<value>s102:2181,s101:2181,s100:2181,</value>
+    </property>
+</configuration>
